+++ R/factory_huxtable.R
@@ -1,22 +1,27 @@
 #' Internal function to build table with `huxtable`
 #'
-#' @inheritParams modelsummary
-#' @param stars_note passed by `modelsummary()`
+#' @inheritParams factory_gt
 #' @keywords internal
 #' @return huxtable object
 factory_huxtable <- function(tab,
-                             title,
-                             stars,
-                             stars_note,
-                             notes,
-                             gof_idx,
-                             output,
+                             align = NULL,
+                             hrule = NULL,
+                             notes = NULL,
+                             output_file = NULL,
+                             output_format = 'huxtable',
+                             title = NULL,
                              ...) {
+
 
     # is huxtable installed?
     if (!requireNamespace('huxtable', quietly = TRUE)) {
         stop("Please install the `huxtable` package.")
     }
+
+    ## warning on align
+    #if (!is.null(align)) {
+        #warning('The `align` argument is not supported yet for huxtable objects. Please file a request or (even better) a code submission on Github if you need that feature.')
+    #}
 
     # huxtable object with header 
     out <- huxtable::hux(tab, add_colnames = TRUE) 
@@ -31,22 +36,20 @@ factory_huxtable <- function(tab,
                                     value = 1) %>%
            huxtable::set_bottom_border(row = nrow(.), 
                                        col = 1:ncol(.),
-                                       value = 1) %>%
-           huxtable::set_bottom_border(row = gof_idx, 
-                                       col = 1:ncol(.),
                                        value = 1)
+
+    if (!is.null(hrule)) {
+        for (pos in hrule) {
+            out <- out %>% 
+                   huxtable::set_bottom_border(row = pos, 
+                                               col = 1:ncol(.),
+                                               value = 1)
+        }
+    }
 
     # title
     if (!is.null(title)) {
         out <- huxtable::set_caption(out, title)
-    }
-
-
-    # stars note
-    stars_note <- make_stars_note(stars)
-    if (!is.null(stars_note)) {
-        out <- huxtable::add_footnote(out,
-                                      text = stars_note)
     }
 
     # user-supplied notes at the bottom of table
@@ -57,19 +60,20 @@ factory_huxtable <- function(tab,
     }
 
     # output
-    ext <- tools::file_ext(output)
-    if (output == 'huxtable') {
+    if (is.null(output_file)) {
         return(out)
-    } else if (ext == 'docx') {
-        huxtable::quick_docx(out, file = output, open = FALSE)
-    } else if (ext == 'pptx') {
-        huxtable::quick_pptx(out, file = output, open = FALSE)
-    } else if (ext %in% c('htm', 'html')) {
-        huxtable::quick_html(out, file = output, open = FALSE)
-    } else if (ext %in% c('rtf')) {
-        huxtable::quick_rtf(out, file = output, open = FALSE)
-    } else if (ext %in% c('tex')) {
-        huxtable::quick_rtf(out, file = output, open = FALSE)
+    } else {
+        if (output_format == 'word') {
+            huxtable::quick_docx(out, file = output_file, open = FALSE)
+        } else if (output_format == 'powerpoint') {
+            huxtable::quick_pptx(out, file = output_file, open = FALSE)
+        } else if (output_format == 'html') {
+            huxtable::quick_html(out, file = output_file, open = FALSE)
+        } else if (output_format == 'rtf') {
+            huxtable::quick_rtf(out, file = output_file, open = FALSE)
+        } else if (output_format == 'latex') {
+            huxtable::quick_rtf(out, file = output_file, open = FALSE)
+        }
     }
 
 }