+++ R/modelsummary.R
@@ -1,7 +1,7 @@
 # https://stackoverflow.com/questions/9439256/how-can-i-handle-r-cmd-check-no-visible-binding-for-global-variable-notes-when#comment20826625_12429344
 # 2012 hadley says "globalVariables is a hideous hack and I will never use it"
 # 2014 hadley updates his own answer with globalVariables as one of "two solutions"
-globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'value', 'p.value', 'std.error', 'statistic', 'stars_note', 'logLik', 'formatBicLL', 'section', 'position'))
+globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'value', 'p.value', 'std.error', 'statistic', 'stars_note', 'logLik', 'formatBicLL', 'section', 'position', 'where', 'ticks', 'statistic1', 'model', 'tmp_grp', 'condition_variable'))
 
 
 #' Beautiful, customizable summaries of statistical models
@@ -11,7 +11,7 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' @param output filename or object type (string)
 #' \itemize{
 #'   \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg. 
-#'   \item Supported object types: "default", "html", "markdown", "latex", "gt", "kableExtra", "huxtable", "flextable".
+#'   \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
 #'   \item When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.
 #' }
 #' @param fmt string which specifies how numeric values will be rounded. This
@@ -49,24 +49,15 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' `omit`. See `modelsummary::gof_map`
 #' @param gof_omit string regular expression. Omits all matching gof statistics from
 #' the table (using `stringr::str_detect`).
-#' @param add_rows a data.frame (or tibble) with the following columns: 
-#' \itemize{
-#'   \item section (character): insert in "middle" or "bottom" section of the table
-#'   \item position (integer): row position in the section
-#'   \item term (character): string to display under coefficient names
-#'   \item one column per model with the same name as that model with the
-#'         values to insert (some models can be omitted). 
-#'   \item See the examples section of this documentation and an example.
-#' }
+#' @param add_rows a data.frame (or tibble) with the same number of columns as
+#' your main table. By default, rows are appended to the bottom of the table.
+#' You can define a "position" attribute of integers to set the row positions.
+#' See examples.
 #' @param title string
 #' @param notes list or vector of notes to append to the bottom of the table.
 #' @param estimate character name of the estimate to display. Must be a column
-#' name in the dataframe produced by `tidy(model)`. In the vast majority of
+#' name in the data.frame produced by `tidy(model)`. In the vast majority of
 #' cases, the default value of this argument should not be changed.
-#' @param filename This argument was deprecated in favor of the `output` argument.
-#' @param subtitle This argument is deprecated. Use `title` or the `tab_header`
-#' @param add_rows_location This argument is deprecated. Use a data.frame as
-#' described in the documentation for the `add_rows` argument.
 #' @param ... all other arguments are passed to the `tidy` method used to
 #' extract estimates from the model. For example, this allows users to set
 #' `exponentiate=TRUE` to exponentiate logistic regression coefficients.
@@ -99,10 +90,11 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' msummary(models, title = gt::md('This is *the* title'))
 #' 
 #' # add_rows: we use `tribble` from the `tibble` package to build a data.frame
-#' # more easily
-#' rows <- tibble::tribble(~term, ~section, ~position, ~Bivariate, ~Multivariate,
-#'                         'Empty row', 'middle', 3, '-', '-',
-#'                         'Another empty row', 'bottom', 1, '?', '?')
+#' # more easily. Then, we assign an attribute to determine each row's position.
+#' rows <- tibble::tribble(~term, ~Bivariate, ~Multivariate,
+#'                         'Empty row', '-', '-',
+#'                         'Another empty row', '?', '?')
+#' attr(rows, 'position') <- c(1, 3)
 #' msummary(models, add_rows = rows)
 #'
 #' # notes at the bottom of the table (here, the second note includes markdown bold characters)
@@ -135,118 +127,64 @@ modelsummary <- function(models,
                          title = NULL,
                          notes = NULL,
                          estimate = 'estimate',
-                         filename = NULL,
-                         subtitle = NULL,
-                         add_rows_location = NULL,
                          ...) {
 
 
-    # sanity check functions are hosted in R/sanity_checks.R
-    # more sanity checks are conducted in modelsummary::extract()
-    sanity_output(output)
-    sanity_title(title)
-    sanity_notes(notes)
-    sanity_filename(filename)
-    sanity_subtitle(subtitle)
-
     # extract estimates and gof
-    dat <- modelsummary::extract(models,
-                                 statistic = statistic,
-                                 statistic_override = statistic_override,
-                                 statistic_vertical = statistic_vertical,
-                                 conf_level = conf_level,
-                                 coef_map = coef_map,
-                                 coef_omit = coef_omit,
-                                 gof_map = gof_map,
-                                 gof_omit = gof_omit,
-                                 stars = stars,
-                                 add_rows = add_rows,
-                                 add_rows_location = add_rows_location,
-                                 fmt = fmt,
-                                 estimate = estimate,
-                                 ...)
+    dat <- extract(models,
+                   statistic = statistic,
+                   statistic_override = statistic_override,
+                   statistic_vertical = statistic_vertical,
+                   conf_level = conf_level,
+                   coef_map = coef_map,
+                   coef_omit = coef_omit,
+                   gof_map = gof_map,
+                   gof_omit = gof_omit,
+                   stars = stars,
+                   fmt = fmt,
+                   estimate = estimate,
+                   ...)
 
     # remove duplicate term labels
     idx <- stringr::str_detect(dat$statistic, 'statistic\\d*$')
     tab <- dat %>%
            dplyr::mutate(term = ifelse(idx, '', term))
 
-    # choose table factory
-    factory_dict <- list('default' = getOption('modelsummary_default', default = 'gt'),
-                         'gt' = 'gt',
-                         'huxtable' = 'huxtable',
-                         'flextable' = 'flextable',
-                         'kableExtra' = 'kableExtra',
-                         'html' = getOption('modelsummary_html', default = 'gt'),
-                         'rtf' = getOption('modelsummary_rtf', default = 'gt'),
-                         'latex' = getOption('modelsummary_latex', default = 'kableExtra'),
-                         'markdown' = 'kableExtra',
-                         'word' =  getOption('modelsummary_word', default = 'flextable'),
-                         'powerpoint' =  getOption('modelsummary_powerpoint', default = 'flextable'),
-                         'png' = getOption('modelsummary_png', default = 'flextable'),
-                         'jpg' = getOption('modelsummary_jpg', default = 'flextable'))
-
-    # sanity check: are user-supplied global options ok?
-    sanity_factory(factory_dict)
-
-    # decide which table factory to use based on user input
-    ext <- tools::file_ext(output)
-
-    # override default if knitting to latex. otherwise gt would break compilation
-    if (output == 'default') {
-        if (knitr::is_latex_output()) {
-            idx <- 'latex'
-        } else {
-            idx <- 'default'
-        }
-
-    # kableExtra produces minimal/readable code
-    } else if (output %in% c('html', 'latex', 'markdown')) { 
-        idx <- 'kableExtra'
-
-    # file extensions to output format
-    } else if (ext %in% c('md', 'Rmd', 'txt')) {
-        idx <- 'markdown'
-    } else if (ext %in% c('tex', 'ltx')) {
-        idx <- 'latex'
-    } else if (ext %in% c('docx', 'doc')) {
-        idx <- 'word'
-    } else if (ext %in% c('pptx', 'ppt')) {
-        idx <- 'powerpoint'
-
-    # human-readable html, latex, markdown 
-    } else if (ext == '') {
-        idx <- output
-
-    # file extension is self-explanatory
-    } else {
-        idx <- ext
+    # measure table
+    hrule <- match('gof', tab$group)
+    if (!is.null(add_rows) && !is.null(attr(add_rows, 'position'))) {
+        hrule <- hrule + sum(attr(add_rows, 'position') < hrule)
     }
 
-    # table factory
-    factory <- factory_dict[[idx]]
-    factory <- list('gt' = factory_gt,
-                    'kableExtra' = factory_kableExtra,
-                    'huxtable' = factory_huxtable,
-                    'flextable' = factory_flextable)[[factory]]
-
-    # clean and measure table
-    gof_idx <- match('gof', tab$group)
+    # clean table
     tab <- tab %>%
            dplyr::select(-statistic, -group) %>%
            # HACK: arbitrary 7 spaces to avoid name conflict
            dplyr::rename(`       ` = term)
 
+
+    # stars
+    if (!isFALSE(stars)) {
+        stars_note <- make_stars_note(stars)
+        if (is.null(notes)) {
+            notes <- stars_note
+        } else {
+            notes <- c(stars_note, notes)
+        }
+    }
+
+    # column alignment left
+    align <- strrep('l', ncol(tab))
+
     # build table
     factory(tab, 
-            title = title,
-            subtitle = subtitle,
-            stars = stars,
-            stars_note = stars_note,
+            align = align,
+            fmt = fmt,
+            hrule = hrule,
             notes = notes,
-            filename = filename,
-            gof_idx = gof_idx,
             output = output,
+            title = title,
+            add_rows = add_rows,
             ...)
       
 }